requirements.txt
aiodns==1.1.1
aiofiles==0.3.1
aiohttp==2.2.5
async-timeout==1.4.0
cchardet==2.1.1
chardet==3.0.4
decorator==4.1.2
flake8==3.4.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mccabe==0.6.1
multidict==3.1.3
peony-twitter==1.0.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycares==2.3.0
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
python-magic==0.4.13
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
yarl==0.12.0
